+++ R/describe_posterior.R
@@ -624,10 +624,15 @@ describe_posterior.BFBayesFactor <- function(posteriors, centrality = "median", 
     compute_bf <- FALSE
   }
 
+  draws <- insight::get_parameters(posteriors)
+  if (rope_range == "default") {
+    rope_range <- rope_range(posteriors)
+  }
+
   # Describe posterior
   out <-
     .describe_posterior(
-      posteriors,
+      draws,
       centrality = centrality,
       dispersion = dispersion,
       ci = ci,
@@ -642,7 +647,7 @@ describe_posterior.BFBayesFactor <- function(posteriors, centrality = "median", 
   # Compute and readd BF a posteriori
   if (compute_bf) {
     tryCatch({
-      out$BF <- as.data.frame(bayesfactor_models(posteriors, ...))[-1, ]$BF
+      out$BF <- as.data.frame(bayesfactor_models(posteriors[1], ...))[-1, ]$BF
     },
     error = function(e) { NULL }
     )