+++ NEWS.md
@@ -13,7 +13,7 @@
 
 ## Bug fixes
 
-* Fixed issue in `describe_posterior()` for multiple-response models.
+* Fixed issue with *default* rope range for `BayesFactor` models.
 * Fixed issue in collinearity-check for `rope()` for models with less than two parameters.
 * Fixed issue in print-method for `mediation()` with `stanmvreg`-models, which displays the wrong name for the response-value.
 * Fixed issue in `effective_sample()` for models with only one parameter.